修改

+++ java/src/main/java/toolgood/flowVision/Flows/SettingFormulaItemWork.java
@@ -9,23 +9,26 @@ public class SettingFormulaItemWork {
     private ProjectWork Project;// { get; set; }
 
     public String Condition;
-    public String Formula ;
+    public String Formula;
 
 
     public Operand EvaluateFormula(FlowEngine engine, InputType inputType) throws Exception {
         mathParser.ProgContext progContext = Project.CreateProgContext(Formula);
-        return engine.EvaluateFormula(progContext, (int)inputType);
+        return engine.EvaluateFormula(progContext, inputType);
     }
 
-    public void Init(ProjectWork work)
-    {
+    public void Init(ProjectWork work) {
         Project = work;
     }
 
 
     public boolean Check(FlowEngine engine) throws Exception {
-        if (Condition==null || Condition.equals("")) { return true; }
-        if (Condition == "1" || Condition.equalsIgnoreCase("true")) { return true; }
+        if (Condition == null || Condition.equals("")) {
+            return true;
+        }
+        if (Condition == "1" || Condition.equalsIgnoreCase("true")) {
+            return true;
+        }
         mathParser.ProgContext progContext = Project.CreateProgContext(Condition);
         return engine.TryEvaluate(progContext, false);
     }
